huggingface_yolo_api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
opencv-python-headless==4.10.0.84
pillow==10.4.0
ultralytics==8.3.197
numpy==2.1.2
pydantic==2.9.2
